refactor: download_req -> __init__.py
