+++ morgan/__init__.py
@@ -2,7 +2,9 @@ from __future__ import annotations
 
 import argparse
 import configparser
+import gzip
 import hashlib
+import json
 import os
 import os.path
 import re
@@ -25,7 +27,6 @@ from morgan.__about__ import __version__
 from morgan.utils import (
     Cache,
     ListExtendingOrderedDict,
-    download_req,
     is_requirement_relevant,
     to_single_dash,
     touch_file,
@@ -832,6 +833,33 @@ def main():  # noqa: C901
         mirror(args)
     elif args.command == "copy_server":
         Mirrorer(args).copy_server()
+
+
+def download_req(index_url: str, req_name: str) -> tuple[dict, str]:
+    # get information about this package from the Simple API in JSON
+    # format as per PEP 691
+    url = index_url.rstrip("/")
+    request = urllib.request.Request(  # noqa: S310
+        f"{url}/{req_name}/",
+        headers={
+            "Accept": "application/vnd.pypi.simple.v1+json",
+            "Accept-Encoding": "gzip",
+        },
+    )
+
+    response_url = ""
+    with urllib.request.urlopen(request) as response:  # noqa: S310
+        # Check if response is gzip-encoded
+        if response.headers.get("Content-Encoding") == "gzip":
+            with gzip.GzipFile(fileobj=response) as gzip_response:
+                data = json.load(gzip_response)
+        else:
+            data = json.load(response)
+        response_url = str(response.url)
+        if data:
+            return data, response_url
+        msg = f"Failed loading metadata: {response}"
+        raise RuntimeError(msg)
 
 
 if __name__ == "__main__":
